requirements.txt
apscheduler>=3.10.4
feedparser>=6.0.11
python-jobspy>=1.1.80
httpx>=0.27.0
beautifulsoup4>=4.12.3
aiosqlite>=0.20.0
python-telegram-bot>=20.7
python-dotenv>=1.0.1
anthropic>=0.40.0
pydantic>=2.6.0
reportlab>=4.1.0
python-docx>=1.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
